+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.5.0-alpha.5 — Unreleased
 - Add links to open file in path completions.
 - Add image preview for image files in path completions.
+- Allow hovering over image/video paths to see preview of image or video.
 
 ## 0.5.0-alpha.4 — April 4, 2024
 - Change update links on paste to generate minimal edit.